% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDConf.R
\name{create_lambda_list_Ridge}
\alias{create_lambda_list_Ridge}
\title{Function for creating a set of candidate lambdas for ridge regression}
\usage{
create_lambda_list_Ridge(X, n_lambdas = 200)
}
\arguments{
\item{X}{is the design matrix}

\item{n_lambdas}{is the number of lambdas we want}
}
\value{
vector of candidate lambdas
}
\description{
Function for creating a set of candidate lambdas for ridge regression
}
