% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDConf.R
\name{CV_select_lambda_Ridge}
\alias{CV_select_lambda_Ridge}
\title{Function for K-fold cross-validation using Ridge regression}
\usage{
CV_select_lambda_Ridge(X, y, n_lambdas = 200, n_folds = 5)
}
\arguments{
\item{X}{is the design matrix}

\item{y}{is the response vector}

\item{n_lambdas}{is the bumber of lambdas we want to test}

\item{n_folds}{is the number of folds}
}
\value{
the average MSE across folds for each lambda and the lambda with the lowest average MSE
}
\description{
Function for K-fold cross-validation using Ridge regression
}
