% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDConf.R
\name{DESLA}
\alias{DESLA}
\title{Function performing de-sparsified LASSO estimation}
\usage{
DESLA(x, y, conf, inv_method = "JM", n_lambdas = 200)
}
\arguments{
\item{x}{is the exposure vector}

\item{y}{is the response vector}

\item{conf}{is the confound matrix}

\item{inv_method}{is the method for estimating the precision matrix, either "JM" (default) (see Javanmard and Montanari (2014)) or "MP" (faster) (see Boot and Nibbering (2017))}

\item{n_lambdas}{is the number of lambdas tested for selection using 5-fold cross-validation}
}
\value{
the p-value and the sign of the coefficient for x
}
\description{
Function performing de-sparsified LASSO estimation
}
