% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDConf.R
\name{HD_OLS}
\alias{HD_OLS}
\title{Function for high-dimensional OLS}
\usage{
HD_OLS(x, y, conf, min_DF = 20)
}
\arguments{
\item{x}{is the exposure vector}

\item{y}{is the response vector}

\item{conf}{is the confound matrix}

\item{min_DF}{is the minimal number of degrees of freedom in the PC-OLS approach}
}
\value{
the p-value and the sign of the coefficient for x
}
\description{
Function for high-dimensional OLS
}
\details{
If the full model has degrees of freedom above min_DF, the function defaults to regular OLS
}
