% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDConf.R
\name{Ridge_PT}
\alias{Ridge_PT}
\title{Function performing permutation testing using ridge regression as described in Hemerik, Thoresen and Finos 2020}
\usage{
Ridge_PT(
  x,
  y,
  conf,
  test_method = "FLH1",
  fit_method = "svd",
  n_perms = 2000,
  n_lambdas = 200
)
}
\arguments{
\item{x}{is the exposure vector}

\item{y}{is the response vector}

\item{conf}{is the confound matrix}

\item{test_method}{is the method used for testing,"FLH1" (default), "FLH2" or "DR"}

\item{fit_method}{is the method used for fitting, "svd" (default) or "glmnet"}

\item{n_perms}{is the number of permutations in the test}

\item{n_lambdas}{is the number of lambdas tested for selection using 5-fold cross-validation}
}
\value{
the p-value and the sign of the correlation between x and y after adjusting for confounds
}
\description{
Function performing permutation testing using ridge regression as described in Hemerik, Thoresen and Finos 2020
}
