% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDConf.R
\name{Ridge_SVD}
\alias{Ridge_SVD}
\title{Function for Ridge regression using SVD}
\usage{
Ridge_SVD(X, y, lambda_list)
}
\arguments{
\item{X}{is the design matrix}

\item{y}{is the response vector}

\item{lambda_list}{is the list of lambdas for which we want to perform Ridge regression}
}
\value{
the ridge coefficients when regressing y on X, using lambda_list as penalties
}
\description{
Function for Ridge regression using SVD
}
